base/requirements.txt
aioredis==1.3.1
aiosmtplib==1.1.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
azure-core==1.9.0
azure-storage-blob==12.6.0
blinker==1.4
boto3==1.16.28
botocore==1.19.28
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
crypto==1.4.1
cryptocode==0.1
cryptography==3.2.1
cssselect==1.1.0
cssutils==1.0.2
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.2
emails==0.6
fakeredis==1.4.5
fastapi==0.63.0
fastapi-mail==0.3.4.2
fastapi-utils==0.2.1
h11==0.11.0
hiredis==1.1.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==1.1.1
isodate==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
lxml==4.6.2
python-magic==0.4.24
MarkupSafe==1.1.1
mongoengine==0.21.0
msrest==0.6.19
Naked==0.1.31
oauthlib==3.1.0
packaging==20.8
pluggy==0.13.1
premailer==3.7.0
py==1.10.0
pyAesCrypt==0.4.3
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.9.9
pydantic==1.7.3
pymongo==3.11.1
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.1
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
s3transfer==0.3.3
shellescape==3.8.1
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.3.0
SQLAlchemy==1.3.20
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.2
XlsxWriter==1.3.7
python-telegram-bot